+++ packages/react/src/number-input/number-input-field.tsx
@@ -3,9 +3,9 @@ import { forwardRef, type ComponentPropsWithoutRef } from 'react'
 import { ark } from '../factory'
 import { useNumberInputContext } from './number-input-context'
 
-export type NumberInputFieldProps = ComponentPropsWithoutRef<typeof ark.input>
+export type NumberInputInputProps = ComponentPropsWithoutRef<typeof ark.input>
 
-export const NumberInputField = forwardRef<HTMLInputElement, NumberInputFieldProps>(
+export const NumberInputInput = forwardRef<HTMLInputElement, NumberInputInputProps>(
   (props, ref) => {
     const { inputProps } = useNumberInputContext()
     const mergedProps = mergeProps(inputProps, props)
@@ -14,4 +14,4 @@ export const NumberInputField = forwardRef<HTMLInputElement, NumberInputFieldPro
   },
 )
 
-NumberInputField.displayName = 'NumberInputField'
+NumberInputInput.displayName = 'NumberInputInput'