+++ packages/solid/src/pin-input/pin-input.stories.tsx
@@ -1,6 +1,6 @@
 import { For } from 'solid-js'
 import type { Meta } from 'storybook-solidjs'
-import { PinInput, PinInputControl, PinInputField, PinInputLabel } from '.'
+import { PinInput, PinInputControl, PinInputInput, PinInputLabel } from '.'
 import './pin-input.css'
 
 const meta: Meta = {
@@ -13,7 +13,7 @@ export const Basic = () => (
   <PinInput placeholder="*" onComplete={(e) => console.log(e)}>
     <PinInputLabel>Label</PinInputLabel>
     <PinInputControl>
-      <For each={[0, 1, 2]}>{(id) => <PinInputField index={id} />}</For>
+      <For each={[0, 1, 2]}>{(id) => <PinInputInput index={id} />}</For>
     </PinInputControl>
   </PinInput>
 )